+++ app/Http/Controllers/TeamController.php
@@ -6,30 +6,12 @@ use App\Models\Team;
 use Illuminate\Http\Request;
 use App\Services\TeamService;
 use App\Services\MessageService;
-use App\Http\Requests\StoreMessageRequest;
 
 class TeamController extends Controller
 {
     public function __construct(MessageService $messageService)
     {
         $this->messageService = $messageService;
-    }
-
-    public function storeMessage(StoreMessageRequest $request, $team)
-    {
-        try {
-            $message = $this->messageService->saveMessage($request->all(), $team);
-
-            return response()->json([
-                'status'  => 'success',
-                'message' => $message,
-            ]);
-        } catch (Exception $e) {
-            return response()->json([
-                'status'  => 'error',
-                'message' => $e->getMessage(),
-            ]);
-        }
     }
 
     public function show(Team $team)
@@ -43,6 +25,7 @@ class TeamController extends Controller
     {
         try {
             $team = $service->createNewTeam($request->all());
+            create_permissions($team);
 
             return response()->json([
                 'status' => 'success',