+++ app/Http/Kernel.php
@@ -51,14 +51,16 @@ class Kernel extends HttpKernel
      * @var array
      */
     protected $routeMiddleware = [
-        'admin'         => \App\Http\Middleware\RedirectIfNotAdmin::class,
-        'auth'          => \Illuminate\Auth\Middleware\Authenticate::class,
-        'auth.basic'    => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
-        'bindings'      => \Illuminate\Routing\Middleware\SubstituteBindings::class,
-        'cache.headers' => \Illuminate\Http\Middleware\SetCacheHeaders::class,
-        'can'           => \Illuminate\Auth\Middleware\Authorize::class,
-        'guest'         => \App\Http\Middleware\RedirectIfAuthenticated::class,
-        'signed'        => \Illuminate\Routing\Middleware\ValidateSignature::class,
-        'throttle'      => \Illuminate\Routing\Middleware\ThrottleRequests::class,
+        'auth'                   => \Illuminate\Auth\Middleware\Authenticate::class,
+        'auth.basic'             => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
+        'bindings'               => \Illuminate\Routing\Middleware\SubstituteBindings::class,
+        'cache.headers'          => \Illuminate\Http\Middleware\SetCacheHeaders::class,
+        'can'                    => \Illuminate\Auth\Middleware\Authorize::class,
+        'guest'                  => \App\Http\Middleware\RedirectIfAuthenticated::class,
+        'granular.permission'    => \App\Http\Middleware\GranularPermissionMiddleware::class,
+        'permission'             => \Spatie\Permission\Middlewares\PermissionMiddleware::class,
+        'role'                   => \Spatie\Permission\Middlewares\RoleMiddleware::class,
+        'signed'                 => \Illuminate\Routing\Middleware\ValidateSignature::class,
+        'throttle'               => \Illuminate\Routing\Middleware\ThrottleRequests::class,
     ];
 }