+++ database/migrations/2017_02_15_045759_create_projects_table.php
@@ -16,7 +16,7 @@ class CreateProjectsTable extends Migration
         Schema::create('projects', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->string('slug');
+            $table->string('slug')->unique();
             $table->string('description');
             $table->integer('office_id')->unsigned()->nullable()->comment('id of office, if any, under which this project operates');
             $table->integer('team_id')->unsigned()->nullable()->comment('id of team, if any, under which this project operates');