+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # Release Notes
 
-## Version no. (Release date)
+## v0.5 (2018-09-16)
 
 ### Added
-- Message ([Issue no.](Issue url))
+- Added `spatie/laravel-permission` package ([issue-152](https://github.com/iluminar/goodwork/issues/152))
+- Added `GranularPermissionMiddleware` ([issue-152](https://github.com/iluminar/goodwork/issues/152))
+- Added `Roles & Permissions` section on Admin page ([issue-152](https://github.com/iluminar/goodwork/issues/152))
+- Added `create_permissions` helper method ([issue-152](https://github.com/iluminar/goodwork/issues/152))
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
-
+- Removed `RedirectIfNotAdmin` middleware
 ### Fixed
 
 ### Security